refactor: issue id array parserをまとめる
