+++ crate-presentation/src/shared.rs
@@ -3,3 +3,4 @@ pub(crate) mod command;
 pub(crate) mod discord_embeds;
 pub(crate) mod discord_utils;
 pub(crate) mod ext;
+pub(crate) mod issue_id_array_parser;
